Moved identical tests to parent class
